Extract save screenshot functionality to the export menu
