+++ src/games/strategy/triplea/ui/menubar/HelpMenu.java
@@ -95,8 +95,8 @@ public class HelpMenu {
           + "Press 'f' to highlight all units you own that have movement left (move phases only).<br>"
           + "Press 'i' or 'v' to popup info on whatever territory and unit your mouse is currently over.<br>"
           + "Press 'u' while mousing over a unit to undo all moves that unit has made (beta).<br>"
-          + "To list specific units from a territory in the Territory panel, drag and drop from the territory on the map to the territory panel.<br>";      final JEditorPane
-          editorPane = new JEditorPane();
+          + "To list specific units from a territory in the Territory panel, drag and drop from the territory on the map to the territory panel.<br>";
+      final JEditorPane editorPane = new JEditorPane();
       editorPane.setEditable(false);
       editorPane.setContentType("text/html");
       editorPane.setText(hints);
@@ -209,13 +209,14 @@ public class HelpMenu {
   }
 
 
+  public static final JEditorPane gameNotesPane = new JEditorPane();
+
   protected void addGameNotesMenu(final JMenu parentMenu) {
     // allow the game developer to write notes that appear in the game
     // displays whatever is in the notes field in html
     final String notesProperty = gameData.getProperties().get("notes", "");
     if (notesProperty != null && notesProperty.trim().length() != 0) {
       final String notes = LocalizeHTML.localizeImgLinksInHTML(notesProperty.trim());
-      JEditorPane gameNotesPane = new JEditorPane();
       gameNotesPane.setEditable(false);
       gameNotesPane.setContentType("text/html");
       gameNotesPane.setText(notes);
